requirements.txt
gitdb2==2.0.6
GitPython==3.0.2
pandas==0.23.0
pygount==1.0.0
numpy==1.14.3
